requirements.txt
aiosqlite==0.20.0
fastapi==0.115.6
httpx==0.27.2
itsdangerous==2.2.0
jinja2==3.1.5
cryptography==44.0.0
python-multipart==0.0.18
python-dotenv==1.0.1
python-telegram-bot==21.6
tzdata==2025.2
uvicorn==0.30.6
